+++ com.blue.auth/src/main/java/com/blue/auth/application/AuthService.java
@@ -8,7 +8,6 @@ import com.blue.auth.domain.UserRepository;
 import com.blue.auth.domain.UserRoleEnum;
 import com.blue.auth.jwt.JwtUtil;
 import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.cglib.core.Local;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
