+++ com.blue.auth/src/main/java/com/blue/auth/domain/UserRepository.java
@@ -1,8 +1,6 @@
 package com.blue.auth.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 
 import java.util.Optional;
